Begin restructuring the code
