+++ src/play.html
@@ -1,0 +1,21 @@
+<!-- This will be the Host view which displays the game board -->
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8" />
+    <title>Robo Rally Online - Play</title>
+  </head>
+  <body>
+    <section id='game-details'></section>
+    <div style="white-space: pre-line" id='board'></div>
+    <section id='message-center'></section>
+    <section id='program-ui'>
+      <section id='program-buttons'></section>
+      <button id='undo-button'>Undo</button>
+      <button id='clear-button'>Clear</button>
+      <button type='button' id='ready-button'>READY</button>
+    </section>
+  <section style="white-space: pre-line" id='program-queue'></section>
+  <section style="white-space: pre-line" id='event-feed'>~~HISTORY~~</section>
+  </body>
+</html>
